Feat(pokemon): move search and paginate to component dir
